--- samples/service/README.md
@@ -1,107 +1,0 @@
-# tomcat 单 host 模式，动态部署多个 web 应用
-tomcat 单host 模式原理介绍详看[这里](https://www.sofastack.tech/projects/sofa-boot/sofa-ark-multi-web-component-deploy/)
-
-# 实验内容
-## 实验应用
-### base
-base 为普通 springboot 改造成的基座，改造内容为在 pom 里增加如下依赖
-```xml
-
-
-<!-- 这里添加动态模块相关依赖 -->
-<dependency>
-    <groupId>com.alipay.sofa.serverless</groupId>
-    <artifactId>sofa-serverless-base-starter</artifactId>
-</dependency>
-<!-- end 动态模块相关依赖 -->
-
-<!-- 这里添加 tomcat 单 host 模式部署多web应用的依赖 -->
-<dependency>
-    <groupId>com.alipay.sofa</groupId>
-    <artifactId>web-ark-plugin</artifactId>
-</dependency>
-<!-- end 单 host 部署的依赖 -->
-```
-
-### biz
-biz 包含两个模块，分别为 biz1 和 biz2, 都是普通 springboot，修改打包插件方式为 sofaArk biz 模块打包方式，打包为 ark biz jar 包，打包插件配置如下：
-```xml
-<!-- 修改打包插件为 sofa-ark biz 打包插件，打包成 ark biz jar -->
-<plugin>
-    <groupId>com.alipay.sofa</groupId>
-    <artifactId>sofa-ark-maven-plugin</artifactId>
-    <version>2.2.3</version>
-    <executions>
-        <execution>
-            <id>default-cli</id>
-            <goals>
-                <goal>repackage</goal>
-            </goals>
-        </execution>
-    </executions>
-    <configuration>
-        <skipArkExecutable>true</skipArkExecutable>
-        <outputDirectory>./target</outputDirectory>
-        <bizName>${bizName}</bizName>
-        <!-- 单host下需更换 web context path -->
-        <webContextPath>${bizName}</webContextPath>
-        <declaredMode>true</declaredMode>
-    </configuration>
-</plugin>
-```
-注意这里将不同 biz 的web context path 修改成不同的值，以此才能成功在一个 tomcat host 里安装多个 web 应用。
-
-
-
-## 实验任务
-1. 执行 mvn clean package -DskipTests
-2. 启动基座应用 base，确保基座启动成功
-3. 执行 curl 命令安装 biz1 和 biz2
-```shell
-curl --location --request POST 'localhost:1238/installBiz' \
---header 'Content-Type: application/json' \
---data '{
-    "bizName": "biz1",
-    "bizVersion": "0.0.1-SNAPSHOT",
-    // local path should start with file://, alse support remote url which can be downloaded
-    "bizUrl": "file:///path/to/sofa-ark-spring-guides/samples/web/tomcat/biz1/target/biz1-web-single-host-0.0.1-SNAPSHOT-ark-biz.jar"
-}'
-```
-
-```shell
-curl --location --request POST 'localhost:1238/installBiz' \
---header 'Content-Type: application/json' \
---data '{
-    "bizName": "biz2",
-    "bizVersion": "0.0.1-SNAPSHOT",
-    // local path should start with file://, alse support remote url which can be downloaded
-    "bizUrl": "file:///path/to/sofa-ark-spring-guides/samples/web/tomcat/biz2/target/biz2-web-single-host-0.0.1-SNAPSHOT-ark-biz.jar"
-}'
-```
-
-如果想验证卸载也可以执行
-```shell
-curl --location --request POST 'localhost:1238/uninstallBiz' \
---header 'Content-Type: application/json' \
---data '{
-    "bizName": "biz1",
-    "bizVersion": "0.0.1-SNAPSHOT"
-}'
-```
-
-4. 发起请求验证
-```shell
-curl http://localhost:8080/biz1
-```
-返回 `hello to /biz1 deploy`
-
-
-```shell
-curl http://localhost:8080/biz2
-```
-返回 `hello to /biz2 deploy`
-
-说明，单host模式应用多次热部署正常。
-
-## 注意事项
-这里主要使用简单应用做验证，如果复杂应用，需要注意模块做好瘦身，基座有的依赖，模块尽可能设置成 provided，尽可能使用基座的依赖。
