--- samples/service/base/src/main/java/com/alipay/sofa/base/impl/SampleServiceImpl.java
@@ -1,14 +1,0 @@
-package com.alipay.sofa.base.impl;
-
-import com.alipay.sofa.model.facade.SampleService;
-import org.springframework.stereotype.Service;
-
-@Service
-public class SampleServiceImpl implements SampleService {
-
-    @Override
-    public String service() {
-        return "A Sample Service";
-    }
-
-}
