--- samples/service/base/src/test/java/com/alipay/sofa/web/base/BaseApplicationTests.java
@@ -1,13 +1,0 @@
-package com.alipay.sofa.web.base;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BaseApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
