--- samples/service/biz/src/test/java/com/alipay/sofa/web/biz1/Biz1ApplicationTests.java
@@ -1,13 +1,0 @@
-package com.alipay.sofa.web.biz1;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class Biz1ApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
