--- samples/service/model/pom.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-	<parent>
-		<groupId>com.alipay.sofa</groupId>
-		<artifactId>sofa-ark-spring-guides</artifactId>
-		<version>0.0.1-SNAPSHOT</version>
-		<relativePath>../../pom.xml</relativePath> <!-- lookup parent from repository -->
-	</parent>
-	<groupId>com.alipay.sofa.service</groupId>
-	<artifactId>model</artifactId>
-	<version>0.0.1-SNAPSHOT</version>
-	<name>model</name>
-	<description>model</description>
-</project>
