--- samples/service/model/src/main/java/com/alipay/sofa/model/facade/AppService.java
@@ -1,7 +1,0 @@
-package com.alipay.sofa.model.facade;
-
-public interface AppService {
-
-    String getAppName();
-
-}
