--- samples/service/model/src/main/java/com/alipay/sofa/model/facade/SampleService.java
@@ -1,10 +1,0 @@
-package com.alipay.sofa.model.facade;
-
-public interface SampleService {
-
-    /**
-     * a simple facade
-     * @return
-     */
-    String service();
-}
