--- samples/service/model/src/main/java/com/alipay/sofa/model/model/Param.java
@@ -1,30 +1,0 @@
-package com.alipay.sofa.model.model;
-
-import java.util.Date;
-
-/**
- * @author: yuanyuan
- * @date: 2023/9/25 8:47 下午
- */
-public class Param {
-
-    private String param1;
-
-    private Date   date;
-
-    public String getParam1() {
-        return param1;
-    }
-
-    public void setParam1(String param1) {
-        this.param1 = param1;
-    }
-
-    public Date getDate() {
-        return date;
-    }
-
-    public void setDate(Date date) {
-        this.date = date;
-    }
-}
