--- samples/service/model/src/main/java/com/alipay/sofa/model/model/People.java
@@ -1,4 +1,0 @@
-package com.alipay.sofa.model.model;
-
-public class People {
-}
