--- samples/service/model/src/main/java/com/alipay/sofa/model/model/Provider.java
@@ -1,7 +1,0 @@
-package com.alipay.sofa.model.model;
-
-
-public interface Provider {
-
-    Result provide(Param param);
-}
