--- samples/service/model/src/main/java/com/alipay/sofa/model/model/Student.java
@@ -1,8 +1,0 @@
-package com.alipay.sofa.model.model;
-
-/**
- * @author: yuanyuan
- * @date: 2023/9/25 8:51 下午
- */
-public class Student extends People {
-}
