+++ samples/service/sample-service-base/base-bootstrap/src/main/java/com/alipay/sofa/base/BaseApplication.java
