+++ samples/service/sample-service-base/base-bootstrap/src/main/java/com/alipay/sofa/base/impl/AppServiceImpl.java
@@ -1,0 +1,17 @@
+package com.alipay.sofa.base.impl;
+
+
+import com.alipay.sofa.base.facade.AppService;
+import org.springframework.stereotype.Service;
+
+/**
+ * @author: yuanyuan
+ * @date: 2023/9/26 2:29 下午
+ */
+@Service
+public class AppServiceImpl implements AppService {
+    @Override
+    public String getAppName() {
+        return null;
+    }
+}
