+++ samples/service/sample-service-base/base-bootstrap/src/main/resources/application.properties
