+++ samples/service/sample-service-base/base-facade/src/main/java/com/alipay/sofa/base/facade/AppService.java
@@ -1,0 +1,10 @@
+package com.alipay.sofa.base.facade;
+
+/**
+ * @author: yuanyuan
+ * @date: 2023/10/16 8:39 下午
+ */
+public interface AppService {
+
+    String getAppName();
+}
