+++ samples/service/sample-service-base/base-facade/src/main/java/com/alipay/sofa/base/facade/SampleService.java
@@ -1,0 +1,10 @@
+package com.alipay.sofa.base.facade;
+
+public interface SampleService {
+
+    /**
+     * a simple facade
+     * @return
+     */
+    String service();
+}
