+++ samples/service/sample-service-biz/biz-bootstrap/pom.xml
@@ -1,0 +1,57 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <parent>
+        <artifactId>sample-service-biz</artifactId>
+        <groupId>com.alipay.sofa.service</groupId>
+        <version>0.0.1-SNAPSHOT</version>
+    </parent>
+    <modelVersion>4.0.0</modelVersion>
+
+    <artifactId>biz-bootstrap</artifactId>
+
+    <properties>
+        <maven.compiler.source>8</maven.compiler.source>
+        <maven.compiler.target>8</maven.compiler.target>
+    </properties>
+
+    <dependencies>
+        <dependency>
+            <groupId>com.alipay.sofa.service</groupId>
+            <artifactId>biz-facade</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.alipay.sofa.service</groupId>
+            <artifactId>base-facade</artifactId>
+        </dependency>
+    </dependencies>
+
+    <build>
+        <plugins>
+            <!-- 修改打包插件为 sofa-ark biz 打包插件，打包成 ark biz jar -->
+            <plugin>
+                <groupId>com.alipay.sofa</groupId>
+                <artifactId>sofa-ark-maven-plugin</artifactId>
+                <version>2.2.3</version>
+                <executions>
+                    <execution>
+                        <id>default-cli</id>
+                        <goals>
+                            <goal>repackage</goal>
+                        </goals>
+                    </execution>
+                </executions>
+                <configuration>
+                    <skipArkExecutable>true</skipArkExecutable>
+                    <outputDirectory>./target</outputDirectory>
+                    <bizName>biz</bizName>
+                    <!-- 单host下需更换 web context path -->
+                    <webContextPath>biz</webContextPath>
+                    <declaredMode>true</declaredMode>
+                </configuration>
+            </plugin>
+        </plugins>
+    </build>
+
+</project>
