+++ samples/service/sample-service-biz/biz-bootstrap/src/main/java/com/alipay/sofa/biz/BizApplication.java
