+++ samples/service/sample-service-biz/biz-bootstrap/src/main/resources/application.properties
