+++ samples/service/sample-service-biz/biz-facade/src/main/java/com/alipay/sofa/biz/facade/Provider.java
@@ -1,0 +1,10 @@
+package com.alipay.sofa.biz.facade;
+
+/**
+ * @author: yuanyuan
+ * @date: 2023/10/16 8:55 下午
+ */
+public interface Provider {
+
+    Result provide(Param param);
+}
