+++ samples/service/sample-service-biz/biz-facade/src/main/java/com/alipay/sofa/biz/facade/Student.java
@@ -1,0 +1,8 @@
+package com.alipay.sofa.biz.facade;
+
+/**
+ * @author: yuanyuan
+ * @date: 2023/10/16 9:34 下午
+ */
+public class Student extends People {
+}
